+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `SearchableList`: The item keys `experimental` and `deprecated` lead to a different rendering.
+
 
 ## [2.0.0-rc.1] - 2020-01-22